Add article architecture with GZIP compression and refine registration ethics

+++ src/main/java/org/mental_management_center/mmc/model/RoleBit.java
@@ -2,19 +2,19 @@ package org.mental_management_center.mmc.model;
 
 public enum RoleBit {
 
-        GUEST((byte) 1),        // 0000 0001 (1)  — Просто зайшов подивитись
-        READER((byte) 2),       // 0000 0010 (2)  — Підписався на контент
-        CLIENT((byte) 4),       // 0000 0100 (4)  — Твій пацієнт/клієнт
-        THERAPIST((byte) 8),    // 0000 1000 (8)  — Колега, фахівець
-        ADMIN((byte) 16);       // 0001 0000 (16) — Ти, господар системи
+        GUEST(1),        // 1 << 0
+        READER(2),       // 1 << 1
+        CLIENT(4),       // 1 << 2
+        THERAPIST(8),    // 1 << 3
+        ADMIN(16);       // 1 << 4
 
-        private final byte mask;
+        private final int mask;
 
-        RoleBit(byte mask) {
+        RoleBit(int mask) {
             this.mask = mask;
         }
 
-        public byte getMask() {
+        public int getMask() {
             return mask;
         }
 }
